fix(formula): test error fix
